+++ apis/locale/locales/en-US.json
@@ -115,6 +115,10 @@
     "value": "Drill-down dimension",
     "comment": "Tooltip for the drill-down icons on the dimension titles in charts. (pde 160210)"
   },
+  "Listbox.Cyclic": {
+    "value": "Cyclic dimension",
+    "comment": "Tooltip for the cyclic icons on the dimension titles in charts. (pde 160210)"
+  },
   "Menu.More": {
     "value": "More",
     "comment": "Menu option to display list of additional actions to choose from (ffd 20200520)"
